Rename mangathr cmd package
