+++ cmd/mangathr/manage/list.go
@@ -2,10 +2,10 @@ package manage
 
 import (
 	"fmt"
-	"github.com/browningluke/mangathrV2/internal/config"
-	"github.com/browningluke/mangathrV2/internal/database"
-	"github.com/browningluke/mangathrV2/internal/logging"
-	"github.com/browningluke/mangathrV2/internal/utils"
+	"github.com/browningluke/mangathr/internal/config"
+	"github.com/browningluke/mangathr/internal/database"
+	"github.com/browningluke/mangathr/internal/logging"
+	"github.com/browningluke/mangathr/internal/utils"
 	"strings"
 )
