+++ cmd/mangathr/manage/manage.go
@@ -3,12 +3,12 @@ package manage
 import (
 	"errors"
 	"fmt"
-	"github.com/browningluke/mangathrV2/internal/config"
-	"github.com/browningluke/mangathrV2/internal/database"
-	"github.com/browningluke/mangathrV2/internal/logging"
-	"github.com/browningluke/mangathrV2/internal/sources"
-	"github.com/browningluke/mangathrV2/internal/ui"
-	"github.com/browningluke/mangathrV2/internal/utils"
+	"github.com/browningluke/mangathr/internal/config"
+	"github.com/browningluke/mangathr/internal/database"
+	"github.com/browningluke/mangathr/internal/logging"
+	"github.com/browningluke/mangathr/internal/sources"
+	"github.com/browningluke/mangathr/internal/ui"
+	"github.com/browningluke/mangathr/internal/utils"
 	"github.com/spf13/cobra"
 )
