+++ cmd/mangathr/update/run.go
@@ -2,12 +2,12 @@ package update
 
 import (
 	"fmt"
-	"github.com/browningluke/mangathrV2/ent"
-	"github.com/browningluke/mangathrV2/internal/database"
-	"github.com/browningluke/mangathrV2/internal/downloader"
-	"github.com/browningluke/mangathrV2/internal/logging"
-	"github.com/browningluke/mangathrV2/internal/sources"
-	"github.com/browningluke/mangathrV2/internal/ui"
+	"github.com/browningluke/mangathr/ent"
+	"github.com/browningluke/mangathr/internal/database"
+	"github.com/browningluke/mangathr/internal/downloader"
+	"github.com/browningluke/mangathr/internal/logging"
+	"github.com/browningluke/mangathr/internal/sources"
+	"github.com/browningluke/mangathr/internal/ui"
 )
 
 // Package-wide accessible driver
